// src/types/Types.d.ts
export type DataRow = {
    year: string,
    category: string,
    winner: string,
    name: string,
    work: string,
    'p:gender': string,
    'p:birthday': string,
    'p:birthplace': string,
    'p:deathday': string,
    'p:popularity': string,
    'p:biography': string,
    'p:knownForDepartment': string,
    'p:profilePath': string,
    'w:posterPath': string,
    'w:genreIds': string,
    'w:popularity': string,
    'w:overview': string
}